Standardize and restructure project
- Use psr4 autoload
- Use composer to describe project and manage dependencies
- Adjust linter style code as psr1 and psr2 rules
- Add helper bin for lint
- Adjust example.php and fix invalid numeric literal
