+++ .gitignore
@@ -1,0 +1,2 @@
+/vendor/
+.php_cs.cache
