+++ composer.json
@@ -1,0 +1,29 @@
+{
+    "name": "nacerns/qnb-payment-integration-class",
+    "description": "QNB Payment Integration is a PHP Class for Integrated Payment via QNB Bank.",
+    "type": "library",
+    "license": "mit",
+    "authors": [
+        {
+            "name": "naciriii",
+            "email": "nnacir1938@hotmail.com"
+        }
+    ],
+    "require": {},
+    "require-dev": {
+        "friendsofphp/php-cs-fixer": "^2.16"
+    },
+    "autoload": {
+        "psr-4": {
+            "Payment\\": "src"
+        }
+    
+    },
+    "scripts": {
+        "lint-check": "./vendor/bin/php-cs-fixer fix ./ --dry-run",
+        "lint": "./vendor/bin/php-cs-fixer fix ./ "
+    }
+    
+    
+
+}
