+++ success.php
@@ -1,3 +1,3 @@
-<?php 
+<?php
 
-	echo "This is a Success Payment Page <br />";
+    echo "This is a Success Payment Page <br />";
